Enemy
Enemy rotation with animation and limited collision in the visual field

+++ TheLastOne/Assets/Scripts/VisualField.cs
@@ -7,7 +7,7 @@ using System.Linq;
 public class VisualField : MonoBehaviour
 {
     public float viewAngle;
-    public int viewRadius;
+    public float viewRadius;
 
     public LayerMask maskTarget;
     public LayerMask maskObstacle;
